requirements.txt
Django==4.2.7
psycopg2-binary>=2.9,<3.0
django-environ>=0.7,<1.0
djangorestframework-simplejwt==5.3.0
djangorestframework==3.14.0
drf-yasg==1.21.7